+++ src/weecfg/__init__.py
@@ -1,12 +1,11 @@
 # coding: utf-8
 #
-#    Copyright (c) 2009-2023 Tom Keffer <tkeffer@gmail.com>
+#    Copyright (c) 2009-2024 Tom Keffer <tkeffer@gmail.com>
 #
 #    See the file LICENSE.txt for your rights.
 #
 """Utilities used by the setup and configure programs"""
 
-import errno
 import importlib
 import os.path
 import pkgutil
@@ -18,8 +17,8 @@ import configobj
 
 import weeutil.config
 import weeutil.weeutil
-from weeutil.weeutil import bcolors
 from weeutil.printer import Printer
+from weeutil.weeutil import bcolors
 
 major_comment_block = ["",
                        "#######################################"
@@ -65,10 +64,8 @@ def find_file(file_path=None, args=None, locations=DEFAULT_LOCATIONS,
         file_path (str|None): A candidate path to the file.
         args (list[str]|None): command-line arguments. If the file cannot be found in file_path,
             then the members of args will be tried.
-        locations (list[str]): A list of directories to be searched. If they do not
-            start with a slash ('/'), then they will be treated as relative to
-            this file (bin/weecfg/__init__.py).
-            Default is ['../..', '/etc/weewx', '/home/weewx'].
+        locations (list[str]): A list of directories to be searched. Default is ['~/weewx-data',
+            '/etc/weewx', '/home/weewx'].
         file_name (str): The name of the file to be found. This is used
             only if the directories must be searched. Default is 'weewx.conf'.
 
@@ -122,10 +119,10 @@ def read_config(config_path, args=None, locations=DEFAULT_LOCATIONS,
         - If it is a relative path, then it is converted to an absolute path by
         prepending the directory where the configuration file was found.
 
-    This version also adds two entries to the returned ConfigObj:
+    This version also adds two possible entries to the returned ConfigObj:
         config_path: Location of the actual configuration file that was used.
-        WEEWX_ROOT_ORIG: The original value of WEEWX_ROOT, or None if there
-            was no original value
+        WEEWX_ROOT_CONFIG: If the original file included a value of WEEWX_ROOT, this is included
+            and set to it. Otherwise, it is not included.
 
     Args:
         config_path (str|None): configuration filename.
@@ -159,15 +156,15 @@ def read_config(config_path, args=None, locations=DEFAULT_LOCATIONS,
 
     # Remember where we found the config file
     config_dict['config_path'] = os.path.realpath(config_path)
-    # Remember the original value for WEEWX_ROOT. In case we need to write this file out, we
-    # can restore it.
-    config_dict['WEEWX_ROOT_ORIG'] = config_dict.get('WEEWX_ROOT')
+    # If there was a value for WEEWX_ROOT in the configuration file, remember it so we can
+    # restore it later.
+    if 'WEEWX_ROOT' in config_dict:
+        config_dict['WEEWX_ROOT_CONFIG'] = config_dict['WEEWX_ROOT']
 
+    # If WEEWX_ROOT is not in the configuration file, supply a default:
     if 'WEEWX_ROOT' not in config_dict:
-        # If missing, set WEEWX_ROOT to the directory the config file is in
         config_dict['WEEWX_ROOT'] = os.path.dirname(config_path)
-    elif config_dict['WEEWX_ROOT'] == '/':
-        config_dict['WEEWX_ROOT'] = '/etc/weewx'
+
     # In case WEEWX_ROOT is a relative path, join it with the location of the config file, then
     # convert it to an absolute path.
     config_dict['WEEWX_ROOT'] = os.path.abspath(os.path.join(os.path.dirname(config_path),
@@ -197,12 +194,24 @@ def save(config_dict, config_path, backup=False):
     write_dict = weeutil.config.deep_copy(config_dict)
     write_dict.pop('config_path', None)
     write_dict.pop('entry_path', None)
-    if write_dict.get('WEEWX_ROOT_ORIG'):
-        # Reset WEEWX_ROOT to what it originally was, then pop WEEWX_ROOT_ORIG
-        write_dict['WEEWX_ROOT'] = write_dict['WEEWX_ROOT_ORIG']
-        write_dict.pop('WEEWX_ROOT_ORIG', None)
 
-    # Check to see if the file exists, and we are supposed to make backup:
+    # If there was a value for WEEWX_ROOT in the config file, restore it
+    if 'WEEWX_ROOT_CONFIG' in write_dict:
+        write_dict['WEEWX_ROOT'] = write_dict['WEEWX_ROOT_CONFIG']
+        #  Add a comment if it doesn't already have one
+        if not write_dict.comments['WEEWX_ROOT']:
+            write_dict.comments['WEEWX_ROOT'] = ['', 'Path to the station data area, relative to '
+                                                     'the configuration file.']
+        del write_dict['WEEWX_ROOT_CONFIG']
+    else:
+        # There was no value for WEEWX_ROOT in the original config file, so delete it.
+        del write_dict['WEEWX_ROOT']
+
+    # If the final path is just '.', get rid of it --- that's the default.
+    if 'WEEWX_ROOT' in write_dict and os.path.normpath(write_dict['WEEWX_ROOT']) == '.':
+        del write_dict['WEEWX_ROOT']
+
+    # Check to see if the file exists, and we are supposed to make a backup:
     if os.path.exists(config_path) and backup:
 
         # Yes. We'll have to back it up.
@@ -637,27 +646,6 @@ def prompt_with_limits(prompt, default=None, low_limit=None, high_limit=None):
 #            Miscellaneous utilities
 # ==============================================================================
 
-def extract_roots(config_dict):
-    """Get the location of the various root directories used by weewx.
-    The extracted paths are *absolute* paths. That is, they are no longer relative to WEEWX_ROOT.
-    """
-    root_dict = dict()
-    root_dict['WEEWX_ROOT'] = config_dict['WEEWX_ROOT']
-    root_dict['USER_DIR'] = os.path.join(config_dict['WEEWX_ROOT'],
-                                         config_dict.get('USER_ROOT', 'bin/user'))
-    root_dict['BIN_DIR'] = os.path.abspath(os.path.join(root_dict['USER_DIR'], '..'))
-    root_dict['EXT_DIR'] = os.path.join(root_dict['USER_DIR'], 'installer')
-
-    # Add SKIN_ROOT if it can be found:
-    try:
-        root_dict['SKIN_DIR'] = os.path.abspath(os.path.join(
-            root_dict['WEEWX_ROOT'],
-            config_dict['StdReport']['SKIN_ROOT']))
-    except KeyError:
-        pass
-
-    return root_dict
-
 
 def extract_tar(filename, target_dir, printer=None):
     """Extract files from a tar archive into a given directory
@@ -707,17 +695,6 @@ def extract_zip(filename, target_dir, printer=None):
 
     del zipfile
     return member_names
-
-
-def mkdir_p(path):
-    """equivalent to 'mkdir -p'"""
-    try:
-        os.makedirs(path)
-    except OSError as e:
-        if e.errno == errno.EEXIST and os.path.isdir(path):
-            pass
-        else:
-            raise
 
 
 def get_extension_installer(extension_installer_dir):