+++ src/weectllib/tests/test_weectllib.py
@@ -1,11 +1,11 @@
 #
-#      Copyright (c) 2023 Tom Keffer <tkeffer@gmail.com>
+#      Copyright (c) 2023-2024 Tom Keffer <tkeffer@gmail.com>
 #
 #      See the file LICENSE.txt for your full rights.
 #
 
-import unittest
 import datetime
+import unittest
 
 from weectllib import parse_dates