+++ src/weeutil/tests/test_config.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 #
-#    Copyright (c) 2020-2023 Tom Keffer <tkeffer@gmail.com>
+#    Copyright (c) 2020-2024 Tom Keffer <tkeffer@gmail.com>
 #
 #    See the file LICENSE.txt for your full rights.
 #