+++ src/weewx/drivers/ultimeter.py
@@ -1,4 +1,4 @@
-# Copyright 2014-2023 Matthew Wall
+# Copyright 2014-2024 Matthew Wall
 # Copyright 2014 Nate Bargmann <n0nb@n0nb.us>
 # See the file LICENSE.txt for your rights.
 #